requirements.txt
# ===============================
# Web / API
# ===============================
fastapi
uvicorn[standard]
python-multipart

# ===============================
# Config / IO
# ===============================
pydantic
pydantic-settings
python-dotenv
httpx

# ===============================
# Agent / Workflow
# ===============================
langchain
langgraph

# ===============================
# UI (Chat)
# ===============================
chainlit

# ===============================
# Vector DB (Local)
# ===============================
faiss-cpu

# ===============================
# Document Parsing
# ===============================
pdfplumber

# ===============================
# LLM / Retry
# ===============================
openai
tenacity

# ===============================
# Logging / JSON
# ===============================
orjson
python-json-logger
rich

# ===============================
# Dev / QA
# ===============================
pytest
pip-tools
